SmartCities_Backend/api/requirements.txt
flask
psycopg2-binary
flask-cors
werkzeug